+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.0.11 - in progress
 - No Donor and Tissue ID validation needed for HCA.
 - Longer description for Q30.
+- Setup GH Pages.
 - Fix bug with loading non-HCA schemas that have an HCA variant.
 
 ## v0.0.10 - 2021-04-21